Refactor mysql workaround code into separate files
